Use temporary kubeconfig file
Write `kubectl config view --flatten --merge` to temporary file and
point `KUBECONFIG` at that. This allows us to muck about with the
context and namespace without impacting the behavior of the regular
`kubectl` command. That is, it means that running `kubecontext` in a
directory with a `.kubecontext` file won't alter the context of your
regular `KUBECONFIG` file.